requirements.txt
torch==2.1.0
torchvision==0.16.0
numpy==1.26.2
Pillow==10.1.0
PyYAML==6.0.1
tqdm==4.66.1
scikit-image==0.22.0
pytest==7.4.3